back-end/requirements.txt
argcomplete==1.10.3
beautifulsoup4==4.8.2
chardet==3.0.4
click==8.1.3
compressed-rtf==1.0.6
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
IMAPClient==2.1.0
joblib==1.2.0
lxml==4.9.2
nltk==3.8.1
numpy==1.24.2
olefile==0.46
pdfminer.six==20191110
Pillow==9.5.0
pycryptodome==3.17
PyPDF2==3.0.1
python-pptx==0.6.21
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.8.1
textract==1.6.5
tqdm==4.65.0
tzdata==2023.3
tzlocal==4.3
xlrd==1.2.0
XlsxWriter==3.0.9
